fastapi/requirements.txt
fastapi
psutil == 5.9.8
python-nmap == 0.7.1
requests == 2.31.0
tqdm == 4.64.1
netifaces == 0.11.0
uvicorn == 0.32.1